feat(lon7.1): establish test sidecar pattern + migrate 2 reference files
Lock in the `#[cfg(test)] #[path = "foo_tests.rs"] mod tests;` pattern
that moves inline test modules into sibling `_tests.rs` files. The
`#[path]` attribute keeps the file as a sibling on disk while the
module remains a child of the source, preserving private-item access
via `use super::*;`.

Reference migrations:

- src/cli/commands/mcp.rs — single-block case (one `mod tests`)
  + src/cli/commands/mcp_tests.rs
- src/ingest/sessions.rs — multi-block case (`mod tests` + `mod decode_tests`)
  + src/ingest/sessions_tests.rs
  + src/ingest/sessions_decode_tests.rs

Documents the convention in CLAUDE.md under "Code Style → Module Layout":
one sidecar per original cfg(test) block, source-side mod name matches
the original, compound cfg gates carry over, `mod test_support;` and
similar non-test helpers stay as regular submodules, monolith policy
already exempts `**/*_tests.*`.

Verified: cargo build + clippy + fmt clean; 1919 → 1919 test count
parity (`cli::commands::mcp::tests::*`, `ingest::sessions::tests::*`,
`ingest::sessions::decode_tests::*` selectors preserved).

Bead: axon_rust-lon7.1
